Move Address to eth_api module
