+++ src/eth/mod.rs
@@ -1,49 +1,11 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use crate::shim::address::Address as FilecoinAddress;
 use cid::{
     multihash::{self, MultihashDigest},
     Cid,
 };
 use std::{fmt, str::FromStr};
-
-const MASKED_ID_PREFIX: [u8; 12] = [0xff, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
-
-#[derive(Default, Clone)]
-pub struct Address(pub ethereum_types::Address);
-
-impl Address {
-    pub fn to_filecoin_address(&self) -> Result<FilecoinAddress, anyhow::Error> {
-        if self.is_masked_id() {
-            // This is a masked ID address.
-            let bytes: [u8; 8] =
-                core::array::from_fn(|i| self.0.as_fixed_bytes()[MASKED_ID_PREFIX.len() + i]);
-            Ok(FilecoinAddress::new_id(u64::from_be_bytes(bytes)))
-        } else {
-            // Otherwise, translate the address into an address controlled by the
-            // Ethereum Address Manager.
-            Ok(FilecoinAddress::new_delegated(
-                FilecoinAddress::ETHEREUM_ACCOUNT_MANAGER_ACTOR.id()?,
-                self.0.as_bytes(),
-            )?)
-        }
-    }
-
-    fn is_masked_id(&self) -> bool {
-        self.0.as_bytes().starts_with(&MASKED_ID_PREFIX)
-    }
-}
-
-impl FromStr for Address {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(Address(
-            ethereum_types::Address::from_str(s).map_err(|e| anyhow::anyhow!("{e}"))?,
-        ))
-    }
-}
 
 #[derive(Default, Clone)]
 pub struct Hash(pub ethereum_types::H256);
