+++ src/lotus_json/eth.rs
@@ -2,23 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 use super::*;
-use crate::eth::{Address, BlockNumberOrHash, Predefined};
-
-impl HasLotusJson for Address {
-    type LotusJson = String;
-
-    fn snapshots() -> Vec<(serde_json::Value, Self)> {
-        vec![(json!("0x0"), Address::default())]
-    }
-
-    fn into_lotus_json(self) -> Self::LotusJson {
-        format!("{:#x}", self.0)
-    }
-
-    fn from_lotus_json(address: Self::LotusJson) -> Self {
-        Address::from_str(&address).unwrap()
-    }
-}
+use crate::eth::{BlockNumberOrHash, Predefined};
 
 impl HasLotusJson for BlockNumberOrHash {
     type LotusJson = String;
