+++ src/rpc/eth_api.rs
@@ -8,7 +8,7 @@ use super::gas_api;
 use crate::blocks::{Tipset, TipsetKeys};
 use crate::chain::{index::ResolveNullTipset, ChainStore};
 use crate::cid_collections::FrozenCidVec;
-use crate::eth::{Address, BlockNumberOrHash, Predefined};
+use crate::eth::{BlockNumberOrHash, Predefined};
 use crate::lotus_json::LotusJson;
 use crate::rpc_api::{data_types::RPCState, eth_api::BigInt as EthBigInt, eth_api::*};
 use crate::shim::{clock::ChainEpoch, state_tree::StateTree};
